store bitcoin_canister endpoint in config
