+++ watchdog/src/remote_api/api_bitaps_com.rs
@@ -18,8 +18,8 @@ pub struct ApiBitapsCom {}
 
 impl ApiBitapsCom {
     /// The host name of the remote API.
-    pub fn host() -> &'static str {
-        "api.bitaps.com"
+    pub fn host() -> String {
+        "api.bitaps.com".to_string()
     }
 
     /// The URL of the remote API.
@@ -30,12 +30,12 @@ impl ApiBitapsCom {
 
     /// Reads the block height from the local storage.
     pub fn get_height() -> Option<BlockHeight> {
-        storage::get(Self::host())
+        storage::get(&Self::host())
     }
 
     /// Stores the block height in the local storage.
     fn set_height(height: BlockHeight) {
-        storage::insert(Self::host(), height)
+        storage::insert(&Self::host(), height)
     }
 
     /// The transform function for the JSON body.
